LLVMLinux: Apply all necessary patches
All patches were found http://git.linuxfoundation.org/?p=llvmlinux.git;a=tree;f=arch/aarch64/patches;h=310490a1c62519fcf358c3dd47f232c3015842a5;hb=HEAD

+++ arch/arm64/include/asm/arch_timer.h
@@ -31,10 +31,12 @@ static inline void arch_timer_reg_write_cp15(int access, int reg, u32 val)
 	if (access == ARCH_TIMER_PHYS_ACCESS) {
 		switch (reg) {
 		case ARCH_TIMER_REG_CTRL:
-			asm volatile("msr cntp_ctl_el0,  %0" : : "r" (val));
+			asm volatile("msr cntp_ctl_el0,  %0"
+				: : "r" ((u64)val));
 			break;
 		case ARCH_TIMER_REG_TVAL:
-			asm volatile("msr cntp_tval_el0, %0" : : "r" (val));
+                       asm volatile("msr cntp_tval_el0, %0"
+                               : : "r" ((u64)val));
 			break;
 		default:
 			BUILD_BUG();
@@ -42,10 +44,14 @@ static inline void arch_timer_reg_write_cp15(int access, int reg, u32 val)
 	} else if (access == ARCH_TIMER_VIRT_ACCESS) {
 		switch (reg) {
 		case ARCH_TIMER_REG_CTRL:
-			asm volatile("msr cntv_ctl_el0,  %0" : : "r" (val));
+                       asm volatile("msr cntv_ctl_el0,  %0"
+                               : : "r" ((u64)val));
+
 			break;
 		case ARCH_TIMER_REG_TVAL:
-			asm volatile("msr cntv_tval_el0, %0" : : "r" (val));
+                       asm volatile("msr cntv_tval_el0, %0"
+                               : : "r" ((u64)val));
+
 			break;
 		default:
 			BUILD_BUG();
@@ -59,7 +65,7 @@ static inline void arch_timer_reg_write_cp15(int access, int reg, u32 val)
 
 static inline u32 arch_timer_reg_read_cp15(int access, int reg)
 {
-	u32 val;
+	u64 val;
 
 	if (access == ARCH_TIMER_PHYS_ACCESS) {
 		switch (reg) {
@@ -87,26 +93,26 @@ static inline u32 arch_timer_reg_read_cp15(int access, int reg)
 		BUILD_BUG();
 	}
 
-	return val;
+	return (u32)val;
 }
 
 static inline u32 arch_timer_get_cntfrq(void)
 {
-	u32 val;
+	u64 val;
 	asm volatile("mrs %0,   cntfrq_el0" : "=r" (val));
-	return val;
+	return (u32)val;
 }
 
 static inline u32 arch_timer_get_cntkctl(void)
 {
-	u32 cntkctl;
+	u64 cntkctl;
 	asm volatile("mrs	%0, cntkctl_el1" : "=r" (cntkctl));
-	return cntkctl;
+	return (u32)cntkctl;
 }
 
 static inline void arch_timer_set_cntkctl(u32 cntkctl)
 {
-	asm volatile("msr	cntkctl_el1, %0" : : "r" (cntkctl));
+	asm volatile("msr	cntkctl_el1, %0" : : "r" ((u64)cntkctl));
 }
 
 static inline void arch_timer_evtstrm_enable(int divider)
